backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
email-validator==2.1.0
requests==2.31.0
python-dotenv==1.0.0
markdown==3.5.1
python-multipart==0.0.6
aiofiles==23.2.1
passlib==1.7.4
bcrypt==4.1.2
python-jose==3.3.0
schedule==1.2.0